requirements.txt
Requirements:

Astor (pip install astor)
PyYAML (pip install PyYAML)